Fix memory usage when not logging
Since all of the text is just being added to a string, it uses way too much memory and makes the script error. This is a temporary fix, maybe a better solution later.
This change also greatly improves performance when not logging, tested on bhop Numismatic a-only wr which is 6 minutes, took 2 seconds to scan.